requirements.txt
asgiref==3.5.2
coverage==5.5
crispy-tailwind==0.2.0
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.0.4
python-decouple==3.4
python-dotenv==0.15.0
pyyaml==5.4.1
validate-email==1.3
whitenoise==5.2.0